Config locations, Dockerfile refactored, mongo config centralized
