+++ basyx.aasenvironment/basyx.aasenvironment.component/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,1 @@
+org.springframework.boot.env.PropertySourceLoader=org.eclipse.digitaltwin.basyx.core.LoggingPropertySourceLoader
