+++ basyx.aasrepository/basyx.aasrepository-backend-mongodb/pom.xml
@@ -13,6 +13,10 @@
 	<dependencies>
 		<dependency>
 			<groupId>org.eclipse.digitaltwin.basyx</groupId>
+			<artifactId>basyx.mongocore</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.eclipse.digitaltwin.basyx</groupId>
 			<artifactId>basyx.aasrepository-core</artifactId>
 		</dependency>
 		<dependency>
