+++ basyx.aasrepository/basyx.aasrepository.component/Dockerfile
@@ -1,6 +1,7 @@
 FROM amazoncorretto:11
 USER nobody
+WORKDIR /application
 ARG JAR_FILE=target/*-exec.jar
-COPY src/main/resources/application.properties /usr/share/config/application.properties
 COPY ${JAR_FILE} basyxExecutable.jar
-ENTRYPOINT ["java","-jar","/basyxExecutable.jar", "--spring.config.location=file:/usr/share/config/application.properties"]
+COPY src/main/resources/application.properties application.properties
+ENTRYPOINT ["java","-jar","basyxExecutable.jar"]
