+++ basyx.aasrepository/basyx.aasrepository.component/application.properties
@@ -1,0 +1,9 @@
+server.port=8081
+basyx.backend = MongoDB
+
+spring.data.mongodb.uri=mongodb://localhost:27017/aas
+
+basyx.aasrepository.feature.mqtt.enabled = true
+mqtt.clientId=AAS-Repo-Example-8081
+mqtt.hostname = localhost
+mqtt.port = 1883
