+++ basyx.aasrepository/basyx.aasrepository.component/src/main/resources/application.properties
