+++ basyx.common/basyx.core/pom.xml
@@ -9,4 +9,15 @@
 	</parent>
 
 	<artifactId>basyx.core</artifactId>
+	
+	<dependencies>
+		<dependency>
+			<groupId>org.springframework</groupId>
+			<artifactId>spring-context</artifactId>
+		</dependency>
+		<dependency>
+			<groupId>org.springframework.boot</groupId>
+			<artifactId>spring-boot-starter</artifactId>
+		</dependency>
+	</dependencies>
 </project>
