+++ basyx.common/basyx.mongocore/src/main/java/org/eclipse/digitaltwin/basyx/common/mongocore/CustomMongoDataAutoConfiguration.java
@@ -24,7 +24,7 @@
  ******************************************************************************/
 
 
-package org.eclipse.digitaltwin.basyx.conceptdescriptionrepository;
+package org.eclipse.digitaltwin.basyx.common.mongocore;
 
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.boot.autoconfigure.data.mongo.MongoDataAutoConfiguration;
