+++ examples/aas-env.properties
@@ -1,0 +1,9 @@
+server.port=8081
+basyx.backend = MongoDB
+
+spring.data.mongodb.uri=mongodb://mongo:27017/aas
+
+#basyx.aasrepository.feature.mqtt.enabled = true
+mqtt.clientId=AAS-Env-Example-8081
+mqtt.hostname = emqx
+mqtt.port = 1883
