+++ examples/aas-repo.properties
@@ -1,0 +1,9 @@
+server.port=8082
+basyx.backend = MongoDB
+
+spring.data.mongodb.uri=mongodb://mongo:27017/aas
+
+#basyx.aasrepository.feature.mqtt.enabled = true
+mqtt.clientId=AAS-Repo-Example-8082
+mqtt.hostname = emqx
+mqtt.port = 1883
